+++ source/src/vm/phc20/memory.cpp
@@ -27,12 +27,12 @@ static const uint8_t key_map[9][8] = {
 #define SET_BANK(s, e, w, r) { \
 	int sb = (s) >> 10, eb = (e) >> 10; \
 	for(int i = sb; i <= eb; i++) { \
-		if((w) == wdmy) { \
+		if(((uintptr_t)w) == (uintptr_t)wdmy) { \
 			wbank[i] = wdmy; \
 		} else { \
 			wbank[i] = (w) + 0x400 * (i - sb); \
 		} \
-		if((r) == rdmy) { \
+		if(((uintptr_t)r) == (uintptr_t)rdmy) { \
 			rbank[i] = rdmy; \
 		} else { \
 			rbank[i] = (r) + 0x400 * (i - sb); \
